// src/services/models/commentResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */
import type { UserResponse } from "./userResponse";

export interface CommentResponse {
  readonly id: number;
  readonly created_at: string;
  readonly updated_at: string;
  readonly walletAddress: string;
  readonly tokenAddress: string;
  readonly content: string;
  readonly createdTime: number;
  readonly userInfo: UserResponse;
}